# openapi.yaml
openapi: 3.0.3
info:
  title: Korvo Atlas API
  description: |
    Verifiable research infrastructure for humans and AI.

    Atlas is an open-source protocol for publishing, validating, and structuring
    research into source-backed, machine-ingestable knowledge objects anchored
    on a blockchain.

    **Authentication:**
    - All `GET` requests are public (no auth required)
    - All write operations (`POST`, `PATCH`, `DELETE`) require a valid API key
    - Pass via header: `Authorization: Bearer <api-key>`
  version: 0.2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Korvo
    url: https://github.com/korvohq/atlas
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://atlas.korvo.io
    description: Production (canonical hub)
security: []
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: Atlas API key (format `atl_...`). Required for write operations.
  schemas:
    Question:
      type: object
      required: [id, text, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        text: { type: string, minLength: 10 }
        context: { type: string }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [open, in_progress, answered, closed] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Source:
      type: object
      required: [id, type, title, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [url, paper, dataset, document, api, testimony, other] }
        title: { type: string }
        url: { type: string, format: uri }
        author: { type: string }
        publishedAt: { type: string, format: date-time }
        retrievedAt: { type: string, format: date-time }
        contentHash: { type: string }
        tags: { type: array, items: { type: string } }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PublishResponse:
      type: object
      properties:
        message: { type: string }
        artifactId: { type: string }
        contentHash: { type: string }
        ipfsCid: { type: string }
        txHash: { type: string }
        chain: { type: string }
        blockNumber: { type: integer }
        publishedAt: { type: string, format: date-time }
    VerifyResponse:
      type: object
      properties:
        verified: { type: boolean }
        contentHashMatch: { type: boolean }
        onChainRecord: { type: object }
        error: { type: string }
...
paths:
...
  /api/questions:
    get:
      summary: List all questions
      tags: [Questions]
      responses:
        '200':
          description: Array of questions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Question' }
    post:
      summary: Create a question
      tags: [Questions]
      security: [{ BearerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, status]
              properties:
                text: { type: string }
                context: { type: string }
                tags: { type: array, items: { type: string } }
                status: { type: string }
                createdBy: { type: string }
      responses:
        '201': { description: Question created }
        '400': { description: Validation error }
        '401': { description: Missing API key }
  /api/questions/{id}:
    get:
      summary: Get question by ID
      tags: [Questions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Question object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '404': { description: Not found }
    patch:
      summary: Update a question
      tags: [Questions]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Updated question }
        '401': { description: Missing API key }
        '404': { description: Not found }
    delete:
      summary: Delete a question
      tags: [Questions]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '401': { description: Missing API key }
        '404': { description: Not found }

  # ── Sources ──
  /api/sources:
    get:
      summary: List all sources
      tags: [Sources]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Source' }
    post:
      summary: Create a source
      tags: [Sources]
      security: [{ BearerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Source created }
        '400': { description: Validation error }
        '401': { description: Missing API key }
  /api/sources/{id}:
    get:
      summary: Get source by ID
      tags: [Sources]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '404': { description: Not found }
    patch:
      summary: Update a source
      tags: [Sources]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Updated }
        '401': { description: Missing API key }
        '404': { description: Not found }
    delete:
      summary: Delete a source
      tags: [Sources]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }

  # ── Claims ──
...
  /api/publish/{artifactId}:
    post:
      summary: Publish an artifact to the blockchain
      tags: [Publish]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: artifactId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                publishedBy: { type: string }
      responses:
        '201':
          description: Successfully published to chain
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublishResponse' }
        '404': { description: Artifact not found }
        '409': { description: Already published }
        '401': { description: Missing API key }
  /api/publish/{artifactId}/verify:
    get:
      summary: Verify artifact against on-chain proof
      tags: [Publish]
      parameters:
        - name: artifactId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Verification passed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResponse' }
        '422': { description: Verification failed }
...